// services/src/app/user/models/ueer-detailed-info.query.response.ts
export interface UserDetailedInfoQueryResponse {
    userId: number;
    name: string;
    email: string;
    mobile: string;
    createdate: Date;
    street: string;
    city: string;
    state: string;
    country: string;
    zipcode: string;
    objective: string;
    companyName: string;
    role: string;
    fromYear: number;
    toYear: number;
    description: string;
    institutionName: string;
    passingYear: number;
    qualification: string;
    university: string;
    percentage: number;
    skillName: string;
    department: string;
    fatherName: string;
    motherName: string;
    dateOfBirth: Date;
    maritalStatus: string;
    languagesKnown: string[];
  }
  